requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.35.3
PyAudio==0.2.14
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
setuptools==69.5.1
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
wheel==0.43.0
